requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
vosk>=0.3.45
TTS>=0.22.0
pyyaml>=6.0
numpy>=1.21.0
pyaudio>=0.2.13
